Struct subspace_node::SecondaryChainCli

pub struct SecondaryChainCli {
    pub run: DomainCli,
    pub base_path: Option<PathBuf>,
    pub chain_spec: ExecutionChainSpec<SystemDomainGenesisConfig>,
}

Fields
run: DomainCli

Run a node.

base_path: Option<PathBuf>

The base path that should be used by the secondary chain.

chain_spec: ExecutionChainSpec<SystemDomainGenesisConfig>

Specification of the secondary chain derived from primary chain spec.

Implementations

impl SecondaryChainCli

pub fn new<'a>(
base_path: Option<PathBuf>,
chain_spec: ExecutionChainSpec<SystemDomainGenesisConfig>,
secondary_chain_args: impl Iterator<Item = &'a String>
) -> (Self, Option<CoreDomainCli>)

Constructs a new instance of SecondaryChainCli.

If no explicit base path for the secondary chain, the default value will be primary_base_path/executor.